// chat-widget-ng5/src/app/chat/enums/actionTypeEnum.ts
export enum ActionType {

    PARKINGLIST = 1,
    REPORTEDINCIDENTlIST = 2,
    WEATHERINFO = 3,
    AIRQUALITYINFO = 4,
    CONNECTAGENT = 5,
    BUSSTOPSLIST = 6,
    BUSDEPOTLIST = 7,
    BUSTRIPPLANNERINFO = 8,
    BUSTRIPDETAILS = 9,
    HELPOPTIONS = 10,
    GARBAGE_SUB_SERVICES = 11,
    WATER_SUB_SERVICES = 12,
    ELECTRICITY_SUB_SERVICES = 13,
    NEARBY_BIN_LOC = 14,
    CONTACT_INFO = 15,
    BUS_ROUTE_NAME_INFO = 16,
    BUS_ROUTE_DETAILS_INFO = 17,
    POLICE_CONTACT_INFO = 18,
    FIRE_CONTACT_INFO = 19,
    AMBULANCE_CONTACT_INFO = 20,
    GREET_SUGGESTION = 21,
    TRANSPORT_SUB_SERVICES = 22,
    BUS_CONTACT_INFO = 23,
    NEAREST_HOSPITALS = 24,
    POLICE_STATION_INFO = 25,
    FIRE_STATION_INFO = 26,
    GARBAGE_CONTACT_INFO = 27,
    AUDIO_ATTACHMENT = 28,
    DOC_ATTACHMENT = 29,
    VIDEO_ATTACHMENT = 30,
    CHAT_BOT_ICON = 31,
    NEAREST_TOURIST_PLACE = 32,
    TWO_WHEELER_PARKINGLIST = 33,
    IMAGE_ATTACHMENT = 34,
    POLICE_VEHICLE_LOCATION = 35,
}

export enum ActionGarbageService {

    REPORT_GABAGE_PROBLEM = 1,
    GARBAGE_PICK_UP_TIMINGS = 2,
    NEAREST_BIN_LOC = 3,
    PAID_GARBAGE_SERVICE = 4,
    GARBAGE_AUTH_CONTACT_DETAILS = 6
}

export enum ActionWaterService {

    REPORT_WATER_PROBLEM = 1,
    WATER_SCHEDULE_TIME = 2,
    NEW_WATER_CONNECTION = 3,
    PAID_WATER_SERVICE = 4,
    WATER_AUTH_CONTACT_DETAILS = 6

}

export enum ActionTransportService {

  BUSSTOPSLIST= 1,
  BUSDEPOTLIST  = 2,
  BUSTRIPDETAILS = 3,
  BUS_ROUTE_INFO= 4,
  BUS_AUTH_CONTACT_DETAILS = 5

}

export enum ActionElectricityService {

    REPORT_ELECTRICITY_PROBLEM = 1,
    NEW_ELECTRICITY_CONNECTION = 2,
    NEARBY_ELECTRICTY_OFFICES = 3,
    ELECTRICITY_BILL_PAYMENT = 4,
    ELECTRICITY_AUTH_CONTACT_DETAILS = 6

}

export enum ActionGreetSuggestionService{

REPORT_INCIDENT = 1,
WEATHER_INFO = 2,
EXPLORE_MORE =3

}